+++ mapsforge-map/src/main/java/org/mapsforge/map/rendertheme/renderinstruction/PathText.java
@@ -50,7 +50,7 @@ public class PathText extends RenderInstruction {
     private final Map<Byte, Paint> fills;
     private float fontSize;
     private int priority;
-    private boolean scaleStrokeWidth = true;
+    private Scale scale = Scale.STROKE;
     private final Paint stroke;
     private final Map<Byte, Paint> strokes;
     private boolean repeat;
@@ -124,8 +124,8 @@ public class PathText extends RenderInstruction {
                 this.rotate = Boolean.parseBoolean(value);
             } else if (PRIORITY.equals(name)) {
                 this.priority = Integer.parseInt(value);
-            } else if (SCALE_STROKE_WIDTH.equals(name)) {
-                this.scaleStrokeWidth = Boolean.parseBoolean(value);
+            } else if (SCALE.equals(name)) {
+                this.scale = scaleFromValue(value);
             } else if (STROKE.equals(name)) {
                 this.stroke.setColor(XmlUtils.getColor(graphicFactory, value));
             } else if (STROKE_WIDTH.equals(name)) {
@@ -187,7 +187,7 @@ public class PathText extends RenderInstruction {
 
     @Override
     public void scaleStrokeWidth(float scaleFactor, byte zoomLevel) {
-        if (!scaleStrokeWidth) {
+        if (this.scale == Scale.NONE) {
             scaleFactor = 1;
         }
         this.dyScaled.put(zoomLevel, this.dy * scaleFactor);